Consolidate critic scripts and refactor optimize function
Consolidate and refactor the reinforcement learning scripts for the project.

* **Add `critic.py`**: Consolidate the functionality of `critic_test1.py`, `critic_test2.py`, and `critic_training.py` into a single script. Refactor the `optimize` function to improve modularity, readability, and robustness. Add comments to explain the purpose of each step in the `optimize` function. Update file paths for saving models and screenshots to be local.
* **Delete redundant files**: Remove `critic_test1.py`, `critic_test2.py`, and `critic_training.py`.
* **Update `README.md`**: Include detailed instructions on the project's structure, purpose, and how to run the consolidated script.

---

For more details, open the [Copilot Workspace session](https://copilot-workspace.githubnext.com/Uggeli/omf_projecti?shareId=XXXX-XXXX-XXXX-XXXX).
